+++ tests/cmaes_test.cpp
@@ -50,7 +50,7 @@ TEST_CASE("CMAESLogisticRegressionTest", "[CMAESTest]")
         responses, testResponses, shuffledResponses);
     LogisticRegression<> lr(shuffledData, shuffledResponses, 0.5);
 
-    CMAES<> cmaes(0, -1, 1, 32, 200, 1e-3);
+    CMAES<> cmaes(0, -20, 20, 32, 200, 1e-3);
     arma::mat coordinates = lr.GetInitialPoint();
     cmaes.Optimize(lr, coordinates);
 
@@ -85,7 +85,7 @@ TEST_CASE("ApproxCMAESLogisticRegressionTest", "[CMAESTest]")
         responses, testResponses, shuffledResponses);
     LogisticRegression<> lr(shuffledData, shuffledResponses, 0.5);
 
-    ApproxCMAES<> cmaes(0, -1, 1, 32, 200, 1e-3);
+    ApproxCMAES<> cmaes(0, -20, 20, 32, 200, 1e-3);
     arma::mat coordinates = lr.GetInitialPoint();
     cmaes.Optimize(lr, coordinates);
 
@@ -120,7 +120,7 @@ TEST_CASE("CMAESLogisticRegressionFMatTest", "[CMAESTest]")
         responses, testResponses, shuffledResponses);
     LogisticRegression<arma::fmat> lr(shuffledData, shuffledResponses, 0.5);
 
-    CMAES<> cmaes(0, -1, 1, 32, 200, 1e-3);
+    CMAES<> cmaes(0, -20, 20, 32, 200, 1e-3);
     arma::fmat coordinates = lr.GetInitialPoint();
     cmaes.Optimize(lr, coordinates);
 
@@ -155,7 +155,7 @@ TEST_CASE("ApproxCMAESLogisticRegressionFMatTest", "[CMAESTest]")
         responses, testResponses, shuffledResponses);
     LogisticRegression<arma::fmat> lr(shuffledData, shuffledResponses, 0.5);
 
-    ApproxCMAES<> cmaes(0, -1, 1, 32, 200, 1e-3);
+    ApproxCMAES<> cmaes(0, -20, 20, 32, 200, 1e-3);
     arma::fmat coordinates = lr.GetInitialPoint();
     cmaes.Optimize(lr, coordinates);
 
@@ -171,4 +171,36 @@ TEST_CASE("ApproxCMAESLogisticRegressionFMatTest", "[CMAESTest]")
   }
 
   REQUIRE(success == true);
+}
+
+/**
+ * Test to check if the boundary transformation
+ * works as expected.
+ */
+TEST_CASE("BoundaryConditionTestFunction", "[CMAESTest]")
+{
+  const size_t trials = 3;
+  for(size_t trail=0; trail < trials; ++trail)
+  {
+    arma::mat data, testData, shuffledData;
+    arma::Row<size_t> responses, testResponses, shuffledResponses;
+
+    LogisticRegressionTestData(data, testData, shuffledData, 
+         responses, testResponses, shuffledResponses);
+    LogisticRegression<> lr(shuffledData, shuffledResponses, 0.5);
+
+    CMAES<> optimizer(0, -10, 10, 32, 200, -1);
+    arma::mat coordinates = lr.GetInitialPoint();
+    optimizer.Optimize(lr, coordinates);
+
+    for (size_t col = 0; col < coordinates.n_cols; col++)
+    {
+      for (size_t row = 0; row < coordinates.n_rows; row++)
+      {
+        bool success = coordinates(row, col) <= 10 && 
+                       coordinates(row, col) >= -10;
+        REQUIRE(success == true);
+      }
+    }
+  }  
 }